ecommerce_project/eComm/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==4.2.27
djangorestframework==3.16.1
idna==3.11
oauthlib==3.3.1
pillow==11.3.0
PyMySQL==1.1.2
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.5
tweepy==4.16.0
typing_extensions==4.15.0
urllib3==2.6.3
